Fix configuration issues, debug failing tests
This is the initial commit of working code. There were a number
of configuration issues with the original source code, the most
prominent being that the ppa repository was no longer available
and python3.6 and all the required dependencies were difficult
to source. In the end I had to compile 3.6 from source and then
find all the relevant dependencies.

In addition to this there were a number of failing tests that had
to be debugged. Some were straightforward but one seemed to stem
from a syntax issue when using boto3; a workaround was found.